module refactoring: putting some things in Servant.Common
