+++ servant/example/greet.hs
@@ -22,7 +22,7 @@ import Servant.API
 import Servant.Client
 import Servant.Docs
 import Servant.Server
-import Servant.Utils.BaseUrl
+import Servant.Common.BaseUrl
 
 -- * Example
