+++ servant/src/Servant/API/Capture.hs
@@ -12,10 +12,10 @@ import GHC.TypeLits
 import Network.Wai
 import Servant.API.Sub
 import Servant.Client
+import Servant.Common.Req
+import Servant.Common.Text
 import Servant.Docs
 import Servant.Server
-import Servant.Utils.Req
-import Servant.Utils.Text
 
 -- * Captures
 data Capture sym a
