+++ servant/src/Servant/API/Delete.hs
@@ -11,10 +11,10 @@ import Data.Typeable
 import Network.HTTP.Types
 import Network.Wai
 import Servant.Client
+import Servant.Common.BaseUrl
+import Servant.Common.Req
 import Servant.Docs
 import Servant.Server
-import Servant.Utils.BaseUrl
-import Servant.Utils.Req
 
 -- | Endpoint for DELETE requests.
 data Delete
