+++ servant/src/Servant/API/Put.hs
@@ -12,10 +12,10 @@ import Data.Typeable
 import Network.HTTP.Types
 import Network.Wai
 import Servant.Client
+import Servant.Common.BaseUrl
+import Servant.Common.Req
 import Servant.Docs
 import Servant.Server
-import Servant.Utils.BaseUrl
-import Servant.Utils.Req
 
 -- | Endpoint for PUT requests.
 data Put a
